requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Ensure compatibility with existing packages
numpy<2,>=1.16.0
scipy<1.12,>=1.4.1